rename docker-util to dockerUtil
