+++ lib/pullImages.js
@@ -4,7 +4,7 @@ const Docker = require('dockerode');
 const { sqldb, sqlLoader } = require('@prairielearn/prairielib');
 
 const logger = require('./logger');
-const dockerUtil = require('./docker-util');
+const dockerUtil = require('./dockerUtil');
 const sql = sqlLoader.loadSqlEquiv(__filename);
 
 module.exports = function(callback) {
